DC-161 Added support for ESS-level local/remote control to site controller (#227)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-161)

## Description

Make site controller aware of an ESS-level maintenance mode or hardware
local control override. This is accomplished through a new register,
`local_mode_signal`, which accepts an integer denoting whether the ESS
is in local control or accepting remote commands from us.

Also includes a fix for parsing of individual bits. Previously we used
generic ints whose size would be platform dependent rather than than
using explicitly sized 64 bit ints. This meant only 16-32 bits would
actually be supported

## How Has This Been Tested?
Pytests. Requires config branch of the same name:
DC-161-Integrating-ESS-maintenance-mode-and-local-remote-mode-status-into-site-controller

+++ site_controller/src/Asset_ESS.cpp
@@ -912,7 +912,7 @@ void Asset_ESS::process_asset() {
         // Check that the bit in the position given by the status value is valid
         // e.g. valid standby states: 4, 5; mask (binary): 110000 (start counting from 0)
         // for status value 4, verify: 110000 & 010000
-        inStandby = standby_status_mask & (1 << internal_status);
+        inStandby = standby_status_mask & (uint64_t(1) << internal_status);
 }
 
 void Asset_ESS::update_asset(void) {
